Same modifications as dev branch

+++ parser/cmds/train.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+import os
 from parser import BiaffineParser, Model
 from parser.utils import Corpus, Embedding, TextDataset, Vocab, collate_fn
 
@@ -32,10 +33,12 @@ class Train(object):
         train = Corpus.load(args.ftrain)
         dev = Corpus.load(args.fdev)
         test = Corpus.load(args.ftest)
-        embed = Embedding.load(args.fembed)
-        vocab = Vocab.from_corpus(corpus=train, min_freq=2)
-        vocab.read_embeddings(embed=embed, unk='unk')
-        torch.save(vocab, args.vocab)
+        if os.path.exists(args.vocab):
+            vocab = torch.load(args.vocab)
+        else:
+            vocab = Vocab.from_corpus(corpus=train, min_freq=2)
+            vocab.read_embeddings(embed=Embedding.load(args.fembed), unk='unk')
+            torch.save(vocab, args.vocab)
         print(vocab)
 
         print("Load the dataset")
